requirements.txt
requests
beautifulsoup4
mtranslate
google-genai
edge-tts
python-slugify
aiohttp
playwright
Pillow